refactor: consolidate error handling tests into test_agent_rearrange
Move tests from test_fix_1464_1465.py into test_agent_rearrange.py
to keep all AgentRearrange tests in one file.
